--- a/tenant/tenant_test.go
+++ b/tenant/tenant_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/grafana/dskit/user"
+
 	"github.com/stretchr/testify/assert"
 )