<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

**What this PR does**:

This improves performance in case there's is only 1 tenant which is apparently 99% of the time. The multi tenant case is kept unchanged.

```
❯ benchstat before.txt after.txt
name                   old time/op    new time/op    delta
TenantID/single-16       84.9ns ± 1%    19.4ns ± 3%   -77.16%  (p=0.008 n=5+5)
TenantID/single#01-16     180ns ± 5%     197ns ±13%    +9.21%  (p=0.032 n=5+5)

name                   old alloc/op   new alloc/op   delta
TenantID/single-16        40.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
TenantID/single#01-16     72.0B ± 0%     72.0B ± 0%      ~     (all equal)

name                   old allocs/op  new allocs/op  delta
TenantID/single-16         2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
TenantID/single#01-16      2.00 ± 0%      2.00 ± 0%      ~     (all equal)
```

Basically saves us 2 allocs per push/query requests.

**Which issue(s) this PR fixes**:

Fixes #<issue number>

**Checklist**
- [x] Tests updated
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]` Don't think I need a changelog entry for that.