@cyriltovena 2024-07-04
good call 
```
❯ benchstat before.txt after.txt
name                old time/op    new time/op    delta
TenantID/single-16    89.9ns ± 3%    19.0ns ± 1%   -78.85%  (p=0.008 n=5+5)
TenantID/multi-16      184ns ± 4%     183ns ± 2%      ~     (p=0.548 n=5+5)

name                old alloc/op   new alloc/op   delta
TenantID/single-16     40.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
TenantID/multi-16      72.0B ± 0%     72.0B ± 0%      ~     (all equal)

name                old allocs/op  new allocs/op  delta
TenantID/single-16      2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
TenantID/multi-16       2.00 ± 0%      2.00 ± 0%      ~     (all equal)

```